User and Scenario model split out to separate files in models/ folder. New line added to each file. #71
