+++ app/models/__init__.py
@@ -1,0 +1,9 @@
+"""
+Models Package.
+Contains database models for the QuantEngine application.
+"""
+
+from app.models.user import User
+from app.models.scenario import Scenario
+
+__all__ = ['User', 'Scenario']
